Backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
pillow==10.3.0
psycopg==3.1.19
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.0
whitenoise==6.6.0
